requirements.txt
transformers
datasets
fastapi
uvicorn
pytesseract
python-docx
pillow
